% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_cols.R
\name{nest_cols}
\alias{nest_cols}
\title{Nest columns together, preserving rows}
\usage{
nest_cols(data, ..., .key = "data")
}
\arguments{
\item{data}{A data frame.}

\item{...}{A selection of columns to nest together. See the
\code{\link[dplyr:select]{dplyr::select()}} documentation for more options.
Cannot include variable used in grouping.}

\item{.key}{The name of the new column, as a string or symbol.}
}
\description{
Nest specified columns together, while preserving rows.
}
\examples{
library(dplyr)

#Place "Sepal" and "Petal" columns into nests

iris \%>\%
 nest_cols(starts_with("Sepal"), .key = "Sepal") \%>\%
 nest_cols(starts_with("Petal"), .key = "Petal")

}
